+++ src/include/migraphx/operators.hpp
@@ -65,6 +65,7 @@
 #include <migraphx/op/gathernd.hpp>
 #include <migraphx/op/get_tuple_elem.hpp>
 #include <migraphx/op/greater.hpp>
+#include <migraphx/op/gridsample.hpp>
 #include <migraphx/op/group.hpp>
 #include <migraphx/op/gru.hpp>
 #include <migraphx/op/identity.hpp>